migration fix

+++ database/migrations/2025_04_12_153252_create_csv_saves_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('csv_saves')) {
+            return;
+        }
         Schema::create('csv_saves', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
